backend/requirements.txt
# Backend requirements for Railway deployment
fastapi==0.115.12
uvicorn[standard]==0.34.1
python-dotenv==1.1.0
supabase==2.15.2
google-generativeai==0.8.5
PyPDF2==3.0.1
python-multipart==0.0.20
flashrank==0.2.10
beautifulsoup4==4.13.4
requests==2.32.3
httpx>=0.27.0
